Added support for bool types

+++ examples/hints_poc/cairo/src/oracle.cairo
@@ -1,11 +1,11 @@
 use starknet::testing::cheatcode;
 #[derive(Drop, Serde)]
 struct RequestUInt32 {
-    n: i64,
+    n: u32,
 }
 #[derive(Drop, Serde)]
 struct ResponseUInt32 {
-    n: i64,
+    n: u32,
 }
 #[generate_trait]
 impl SqrtOracle of SqrtOracleTrait {
